Seperating viz from structures.
