+++ beamz/design/structures.py
@@ -4,6 +4,7 @@ from matplotlib.path import Path
 import random
 import numpy as np
 from beamz.design.materials import Material
+from beamz import viz as viz
 from beamz.const import µm, EPS_0, MU_0
 from beamz.design.sources import ModeSource, GaussianSource
 from beamz.design.monitors import Monitor
@@ -506,1015 +507,20 @@ class Design:
         self.structures.append(top_pml)
 
     def show(self, unify_structures=True):
-        """Display the design visually using 2D matplotlib or 3D plotly."""
-        is_truly_3d = self._determine_if_3d()
-        if is_truly_3d: self.show_3d(unify_structures)
-        else: self.show_2d(unify_structures)
+        """Display the design visually using 2D matplotlib or 3D plotly (delegated to beamz.viz)."""
+        return viz.show_design(self, unify_structures)
     
     def _determine_if_3d(self):
-        """Determine if the design should be visualized in 3D based on structure properties."""
-        # Check design-level depth
-        if self.depth and self.depth > 0:
-            # If any structure has non-zero depth or z-position, it's 3D
-            for structure in self.structures:
-                # Skip PML structures from consideration
-                if hasattr(structure, 'is_pml') and structure.is_pml: continue
-                # Check for non-zero depth
-                if hasattr(structure, 'depth') and structure.depth and structure.depth > 0: return True
-                # Check for non-zero z position
-                if hasattr(structure, 'z') and structure.z and structure.z != 0: return True
-                # Check position z-coordinate 
-                if hasattr(structure, 'position') and len(structure.position) > 2 and structure.position[2] != 0: return True
-                # Check if any vertex has non-zero z coordinate
-                if hasattr(structure, 'vertices') and structure.vertices:
-                    for vertex in structure.vertices:
-                        if len(vertex) > 2 and vertex[2] != 0: return True
-        return False
+        """Determine if the design should be visualized in 3D (delegated to beamz.viz)."""
+        return viz.determine_if_3d(self)
     
     def show_2d(self, unify_structures=True):
-        """Display the design using 2D matplotlib visualization."""
-        # Determine appropriate SI unit and scale
-        max_dim = max(self.width, self.height)
-        scale, unit = get_si_scale_and_label(max_dim)
-        # Calculate figure size based on domain dimensions
-        aspect_ratio = self.width / self.height
-        base_size = 5
-        if aspect_ratio > 1: figsize = (base_size * aspect_ratio, base_size)
-        else: figsize = (base_size, base_size / aspect_ratio)
-        # Decide which structures to plot without mutating the original design
-        if unify_structures:
-            tmp_design = self.copy()
-            tmp_design.unify_polygons()
-            structures_to_plot = tmp_design.structures
-        else:
-            structures_to_plot = self.structures
-        # Create a single figure for all structures
-        fig, ax = plt.subplots(figsize=figsize)
-        ax.set_aspect('equal')
-        # Now plot each structure
-        for structure in structures_to_plot:
-            # Use dashed lines for PML regions
-            if hasattr(structure, 'is_pml') and structure.is_pml:
-                structure.add_to_plot(ax, edgecolor=self.border_color, linestyle='--', facecolor='none', alpha=0.5)
-            else: structure.add_to_plot(ax)
-        
-        # Plot PML boundaries explicitly with dashed lines
-        for boundary in self.boundaries:
-            if hasattr(boundary, 'add_to_plot'):
-                boundary.add_to_plot(ax, edgecolor=self.border_color, linestyle='--', facecolor='none', alpha=0.5)
-        
-        # Set proper limits, title and label, and ensure the full design is visible
-        ax.set_title('Design Layout')
-        ax.set_xlabel(f'X ({unit})')
-        ax.set_ylabel(f'Y ({unit})')
-        ax.set_xlim(0, self.width)
-        ax.set_ylim(0, self.height)
-        # Update tick labels with scaled values
-        ax.xaxis.set_major_formatter(lambda x, pos: f'{x*scale:.1f}')
-        ax.yaxis.set_major_formatter(lambda x, pos: f'{x*scale:.1f}')
-        # Adjust layout for clean appearance
-        plt.tight_layout()
-        plt.show()
+        """Display the design using 2D matplotlib (delegated to beamz.viz)."""
+        return viz.show_design_2d(self, unify_structures)
     
     def show_3d(self, unify_structures=True, max_vertices_for_unification=50):
-        """Display the design using 3D plotly visualization."""
-        try:
-            import plotly.graph_objects as go
-            import plotly.figure_factory as ff
-            from plotly.subplots import make_subplots
-        except ImportError:
-            display_status("Plotly is required for 3D visualization. Install with: pip install plotly", "error")
-            display_status("Falling back to 2D visualization...", "warning")
-            self.show_2d(unify_structures)
-            return
-        # Determine appropriate SI unit and scale
-        max_dim = max(self.width, self.height, self.depth if self.depth else 0)
-        scale, unit = get_si_scale_and_label(max_dim)
-        # For 3D visualization, be more selective about unification
-        if unify_structures:
-            # Check if we have complex structures that might not unify well
-            complex_structures = 0
-            total_vertices = 0
-            for structure in self.structures:
-                if hasattr(structure, 'vertices') and structure.vertices:
-                    vertices_count = len(structure.vertices)
-                    total_vertices += vertices_count
-                    if vertices_count > max_vertices_for_unification: complex_structures += 1
-            # Disable unification if we have too many complex structures
-            if complex_structures > 2 or total_vertices > 200:
-                display_status(f"Disabling polygon unification for 3D (too complex: {complex_structures} complex structures, \
-                                {total_vertices} total vertices)", "warning")
-                unify_structures = False
-            else: self.unify_polygons()
-        
-        # Create 3D figure with modern styling
-        fig = go.Figure()
-        # Default depth for 2D structures in 3D view
-        default_depth = self.depth if self.depth else min(self.width, self.height) * 0.1
-        
-        # Modern color palette for different materials
-        modern_colors = [
-            '#1f77b4',  # Blue
-            '#ff7f0e',  # Orange  
-            '#2ca02c',  # Green
-            '#d62728',  # Red
-            '#9467bd',  # Purple
-            '#8c564b',  # Brown
-            '#e377c2',  # Pink
-            '#7f7f7f',  # Gray
-            '#bcbd22',  # Olive
-            '#17becf'   # Cyan
-        ]
-        
-        # Track materials for consistent coloring
-        material_colors = {}
-        color_index = 0
-        # Process each structure
-        for structure in self.structures:
-            # Skip PML structures for now (they would clutter the 3D view)
-            if hasattr(structure, 'is_pml') and structure.is_pml: continue
-                
-            # Handle Monitor objects specially
-            if isinstance(structure, Monitor):
-                self._add_monitor_to_3d_plot(fig, structure, scale, unit)
-                continue
-                
-            # Handle ModeSource objects specially
-            if isinstance(structure, ModeSource):
-                self._add_mode_source_to_3d_plot(fig, structure, scale, unit)
-                continue
-                
-            # Handle GaussianSource objects specially
-            if isinstance(structure, GaussianSource):
-                self._add_gaussian_source_to_3d_plot(fig, structure, scale, unit)
-                continue
-                
-            # Get structure depth
-            struct_depth = getattr(structure, 'depth', default_depth)
-            struct_z = getattr(structure, 'z', 0)
-            # Add 3D representation based on structure type
-            mesh_data = self._structure_to_3d_mesh(structure, struct_depth, struct_z)
-            if mesh_data:
-                x, y, z = mesh_data['vertices']
-                i, j, k = mesh_data['faces']
-                
-                # Get material properties for transparency determination
-                material_permittivity = 1.0  # Default air/vacuum
-                if hasattr(structure, 'material') and structure.material:
-                    material_permittivity = getattr(structure.material, 'permittivity', 1.0)
-                
-                # Assign color based on material properties for consistency
-                material_key = None
-                if hasattr(structure, 'material') and structure.material:
-                    material_key = (
-                        getattr(structure.material, 'permittivity', 1.0),
-                        getattr(structure.material, 'permeability', 1.0),
-                        getattr(structure.material, 'conductivity', 0.0)
-                    )
-                
-                if material_key not in material_colors:
-                    if hasattr(structure, 'color') and structure.color and structure.color != 'none':
-                        material_colors[material_key] = structure.color
-                    else:
-                        material_colors[material_key] = modern_colors[color_index % len(modern_colors)]
-                        color_index += 1
-                
-                color = material_colors[material_key]
-                
-                # Determine transparency based on material properties
-                # Materials close to air/vacuum (permittivity ≈ 1.0) are fully transparent
-                # All other materials are fully opaque
-                is_air_like = abs(material_permittivity - 1.0) < 0.1  # Within 10% of air/vacuum
-                
-                # Convert hex color to rgba with appropriate transparency
-                if color.startswith('#'):
-                    r, g, b = int(color[1:3], 16), int(color[3:5], 16), int(color[5:7], 16)
-                    if is_air_like:
-                        face_color = f"rgba({r},{g},{b},0.0)"  # Fully transparent for air/vacuum
-                        opacity = 0.0
-                    else:
-                        face_color = f"rgba({r},{g},{b},1.0)"  # Fully opaque for all other materials
-                        opacity = 1.0
-                else:
-                    face_color = color
-                    opacity = 0.0 if is_air_like else 1.0
-                
-                # Add material name to hover info if available
-                hovertext = f"{structure.__class__.__name__}"
-                if hasattr(structure, 'material') and structure.material:
-                    if hasattr(structure.material, 'name'):
-                        hovertext += f"<br>Material: {structure.material.name}"
-                    if hasattr(structure.material, 'permittivity'):
-                        hovertext += f"<br>εᵣ = {structure.material.permittivity:.1f}"
-                    if hasattr(structure.material, 'permeability') and structure.material.permeability != 1.0:
-                        hovertext += f"<br>μᵣ = {structure.material.permeability:.1f}"
-                    if hasattr(structure.material, 'conductivity') and structure.material.conductivity != 0.0:
-                        hovertext += f"<br>σ = {structure.material.conductivity:.2e} S/m"
-                
-                fig.add_trace(go.Mesh3d(
-                    x=x, y=y, z=z,
-                    i=i, j=j, k=k,
-                    color=face_color,  # Use 'color' for uniform coloring
-                    opacity=opacity,  # Fully transparent for air/vacuum, fully opaque for others
-                    name=hovertext,
-                    showscale=True,
-                    hovertemplate=hovertext + "<extra></extra>",
-                    # Prominent black outlines for ALL shapes (including transparent ones)
-                    contour=dict(
-                        show=True,
-                        color="black",  # Always black outlines
-                        width=5  # Thick lines for better visibility
-                    ),
-                    # Flat shading - minimal lighting for clean appearance
-                    lighting=dict(
-                        ambient=0.5,    # High ambient for flat appearance
-                        diffuse=0.5,    # Low diffuse for minimal shadows
-                        fresnel=0.0,    # No fresnel effects
-                        specular=0.5,   # No specular highlights  
-                        roughness=1.0   # Maximum roughness for flat appearance
-                    ),
-                    # Simple light position
-                    lightposition=dict(
-                        x=0, y=50, z=100  # Far away light for even illumination
-                    ),
-                    # Force flat shading
-                    flatshading=True
-                ))
-        
-        # Modern layout with better styling
-        fig.update_layout(
-            scene=dict(
-                xaxis=dict(
-                    title=dict(text=f'X ({unit})', font=dict(size=14, color='#34495e')),
-                    range=[0, self.width],
-                    showgrid=True,
-                    gridcolor='rgba(128,128,128,0.3)',
-                    showbackground=True,
-                    backgroundcolor='rgba(248,249,250,0.8)',
-                    tickmode='array',
-                    tickvals=np.linspace(0, self.width, 6),
-                    ticktext=[f'{val*scale:.1f}' for val in np.linspace(0, self.width, 6)],
-                    tickfont=dict(size=11, color='#34495e')
-                ),
-                yaxis=dict(
-                    title=dict(text=f'Y ({unit})', font=dict(size=14, color='#34495e')),
-                    range=[0, self.height],
-                    showgrid=True,
-                    gridcolor='rgba(128,128,128,0.3)',
-                    showbackground=True,
-                    backgroundcolor='rgba(248,249,250,0.8)',
-                    tickmode='array',
-                    tickvals=np.linspace(0, self.height, 6),
-                    ticktext=[f'{val*scale:.1f}' for val in np.linspace(0, self.height, 6)],
-                    tickfont=dict(size=11, color='#34495e')
-                ),
-                zaxis=dict(
-                    title=dict(text=f'Z ({unit})', font=dict(size=14, color='#34495e')),
-                    range=[0, self.depth if self.depth else default_depth],
-                    showgrid=True,
-                    gridcolor='rgba(128,128,128,0.3)',
-                    showbackground=True,
-                    backgroundcolor='rgba(248,249,250,0.8)',
-                    tickmode='array',
-                    tickvals=np.linspace(0, self.depth if self.depth else default_depth, 6),
-                    ticktext=[f'{val*scale:.1f}' for val in np.linspace(0, self.depth if self.depth else default_depth, 6)],
-                    tickfont=dict(size=11, color='#34495e')
-                ),
-                aspectmode='manual',
-                aspectratio=dict(
-                    x=1,
-                    y=self.height/self.width if self.width > 0 else 1,
-                    z=(self.depth if self.depth else default_depth)/self.width if self.width > 0 else 1
-                ),
-                # Modern camera angle for better initial view
-                camera=dict(
-                    eye=dict(x=1.5, y=1.5, z=1.2),
-                    center=dict(x=0, y=0, z=0),
-                    up=dict(x=0, y=0, z=1)
-                )
-            ),
-            # Modern layout styling
-            width=900,
-            height=700,
-            margin=dict(l=60, r=60, t=80, b=60),
-            paper_bgcolor='white',
-            plot_bgcolor='white',
-            font=dict(family="Arial, sans-serif", size=12, color='#2c3e50'),
-            # Enhanced legend
-            showlegend=True,
-            legend=dict(
-                orientation="v",
-                yanchor="top",
-                y=1,
-                xanchor="left", 
-                x=1.02,
-                bgcolor="rgba(255,255,255,0.8)",
-                bordercolor="rgba(0,0,0,0.2)",
-                borderwidth=1,
-                font=dict(size=10)
-            )
-        )
-        
-        # Add subtle shadow effect with a ground plane if structures are elevated
-        if any(getattr(s, 'z', 0) > 0 for s in self.structures if not (hasattr(s, 'is_pml') and s.is_pml)):
-            # Create a subtle ground plane
-            ground_x = [0, self.width, self.width, 0]
-            ground_y = [0, 0, self.height, self.height]
-            ground_z = [0, 0, 0, 0]
-            fig.add_trace(go.Mesh3d(
-                x=ground_x + ground_x,  # Duplicate for thickness
-                y=ground_y + ground_y,
-                z=ground_z + [-default_depth*0.05]*4,  # Thin ground plane
-                i=[0, 0, 4, 4, 0, 1, 2, 3],
-                j=[1, 3, 5, 7, 4, 5, 6, 7], 
-                k=[2, 2, 6, 6, 1, 2, 3, 0],
-                color='rgba(220,220,220,0.3)',  # Use 'color' instead of 'facecolor'
-                name="Ground Plane",
-                showlegend=False,
-                hoverinfo='skip',
-                # Consistent flat shading for ground plane
-                lighting=dict(
-                    ambient=0.8,
-                    diffuse=0.2,
-                    fresnel=0.0,
-                    specular=0.0,
-                    roughness=1.0
-                ),
-                flatshading=True,
-                # No outline for ground plane (to keep it subtle)
-                contour=dict(show=True, color="black", width=5)
-            ))
-        
-        fig.show()
-    
-    def _add_monitor_to_3d_plot(self, fig, monitor, scale, unit):
-        """Add a monitor plane to the 3D plot."""
-        try:
-            import plotly.graph_objects as go
-        except ImportError:
-            return  # Skip if plotly not available
-            
-        if not hasattr(monitor, 'vertices') or not monitor.vertices:
-            return  # Skip if no vertices
-            
-        # Get monitor vertices
-        vertices = monitor.vertices
-        if len(vertices) < 3:
-            return  # Need at least 3 vertices
-            
-        # Extract coordinates
-        x_coords = [v[0] for v in vertices]
-        y_coords = [v[1] for v in vertices]
-        z_coords = [v[2] for v in vertices]
-        
-        # Create a simple triangulation for the monitor plane (it's a rectangle)
-        if len(vertices) == 4:
-            # Rectangle: two triangles
-            faces_i = [0, 0]
-            faces_j = [1, 2]
-            faces_k = [2, 3]
-        else:
-            # For other shapes, use simple fan triangulation
-            faces_i, faces_j, faces_k = [], [], []
-            for i in range(1, len(vertices) - 1):
-                faces_i.append(0)
-                faces_j.append(i)
-                faces_k.append(i + 1)
-        
-        # Create hover text
-        hovertext = f"Monitor ({monitor.monitor_type})"
-        if hasattr(monitor, 'size'):
-            hovertext += f"<br>Size: {monitor.size[0]*scale:.2f} x {monitor.size[1]*scale:.2f} {unit}"
-        if hasattr(monitor, 'plane_normal'):
-            hovertext += f"<br>Normal: {monitor.plane_normal}"
-        if hasattr(monitor, 'plane_position'):
-            hovertext += f"<br>Position: {monitor.plane_position*scale:.2f} {unit}"
-        
-        # Add the monitor as a semi-transparent mesh
-        fig.add_trace(go.Mesh3d(
-            x=x_coords, y=y_coords, z=z_coords,
-            i=faces_i, j=faces_j, k=faces_k,
-            color='rgba(255,255,0,0.6)',  # Semi-transparent yellow
-            opacity=0.75,
-            name="Monitor",
-            hovertemplate=hovertext + "<extra></extra>",
-            # Prominent outline for visibility
-            contour=dict(
-                show=True,
-                color="black",
-                width=8  # Thick lines for monitor visibility
-            ),
-            # Flat shading for clean appearance
-            lighting=dict(
-                ambient=0.8,
-                diffuse=0.2,
-                fresnel=0.0,
-                specular=0.0,
-                roughness=1.0
-            ),
-            flatshading=True,
-            showlegend=True
-        ))
-    
-    def _add_mode_source_to_3d_plot(self, fig, source, scale, unit):
-        """Add a mode source plane to the 3D plot."""
-        try:
-            import plotly.graph_objects as go
-        except ImportError:
-            return  # Skip if plotly not available
-            
-        # Use the new position-based approach if available
-        if hasattr(source, 'width') and hasattr(source, 'height') and hasattr(source, 'orientation'):
-            # New position-based approach
-            center = source.position
-            width = source.width
-            height = source.height if source.height > 0 else source.wavelength * 0.5  # Default height
-            orientation = getattr(source, 'orientation', 'yz')
-            
-            # Create vertices based on orientation
-            if orientation == "yz":
-                # Cross-section in yz plane
-                vertices = [
-                    (center[0], center[1] - width/2, center[2] - height/2),
-                    (center[0], center[1] + width/2, center[2] - height/2),
-                    (center[0], center[1] + width/2, center[2] + height/2),
-                    (center[0], center[1] - width/2, center[2] + height/2)
-                ]
-            elif orientation == "xz":
-                # Cross-section in xz plane
-                vertices = [
-                    (center[0] - width/2, center[1], center[2] - height/2),
-                    (center[0] + width/2, center[1], center[2] - height/2),
-                    (center[0] + width/2, center[1], center[2] + height/2),
-                    (center[0] - width/2, center[1], center[2] + height/2)
-                ]
-            else:  # orientation == "xy"
-                # Cross-section in xy plane
-                vertices = [
-                    (center[0] - width/2, center[1] - height/2, center[2]),
-                    (center[0] + width/2, center[1] - height/2, center[2]),
-                    (center[0] + width/2, center[1] + height/2, center[2]),
-                    (center[0] - width/2, center[1] + height/2, center[2])
-                ]
-        else:
-            # Legacy approach using start and end points
-            start = source.start
-            end = source.end
-            
-            # Calculate the line vector and its length
-            line_vec = np.array([end[0] - start[0], end[1] - start[1], end[2] - start[2]])
-            line_length = np.linalg.norm(line_vec)
-            
-            if line_length == 0:
-                # If start and end are the same, create a small square plane
-                center = start
-                # Create a small plane in the xy direction
-                plane_size = source.wavelength * 0.5  # Half wavelength size
-                vertices = [
-                    (center[0] - plane_size/2, center[1] - plane_size/2, center[2]),
-                    (center[0] + plane_size/2, center[1] - plane_size/2, center[2]),
-                    (center[0] + plane_size/2, center[1] + plane_size/2, center[2]),
-                    (center[0] - plane_size/2, center[1] + plane_size/2, center[2])
-                ]
-            else:
-                # Create a plane perpendicular to the line
-                line_unit = line_vec / line_length
-                
-                # Find two perpendicular vectors to create the plane
-                # Choose an arbitrary vector not parallel to line_unit
-                if abs(line_unit[2]) < 0.9:
-                    temp_vec = np.array([0, 0, 1])
-                else:
-                    temp_vec = np.array([1, 0, 0])
-                
-                # Create two orthogonal vectors in the plane
-                perp1 = np.cross(line_unit, temp_vec)
-                perp1 = perp1 / np.linalg.norm(perp1)
-                perp2 = np.cross(line_unit, perp1)
-                perp2 = perp2 / np.linalg.norm(perp2)
-                
-                # Size the plane based on the line length or wavelength
-                plane_size = max(line_length, source.wavelength * 0.5)
-                
-                # Create vertices for a rectangular plane
-                center = np.array([(start[0] + end[0])/2, (start[1] + end[1])/2, (start[2] + end[2])/2])
-                
-                vertices = [
-                    center - perp1 * plane_size/2 - perp2 * plane_size/2,
-                    center + perp1 * plane_size/2 - perp2 * plane_size/2,
-                    center + perp1 * plane_size/2 + perp2 * plane_size/2,
-                    center - perp1 * plane_size/2 + perp2 * plane_size/2
-                ]
-                
-                # Convert to tuples
-                vertices = [(v[0], v[1], v[2]) for v in vertices]
-        
-        # Extract coordinates
-        x_coords = [v[0] for v in vertices]
-        y_coords = [v[1] for v in vertices]
-        z_coords = [v[2] for v in vertices]
-        
-        # Create triangulation for the plane (rectangle = 2 triangles)
-        faces_i = [0, 0]
-        faces_j = [1, 2]
-        faces_k = [2, 3]
-        
-        # Create hover text
-        hovertext = f"ModeSource"
-        hovertext += f"<br>Wavelength: {source.wavelength*scale*1e6:.0f} nm"
-        hovertext += f"<br>Direction: {source.direction}"
-        hovertext += f"<br>Modes: {source.num_modes}"
-        if hasattr(source, 'effective_indices') and len(source.effective_indices) > 0:
-            hovertext += f"<br>n_eff: {source.effective_indices[0].real:.3f}"
-        
-        # Add the source as a semi-transparent mesh
-        fig.add_trace(go.Mesh3d(
-            x=x_coords, y=y_coords, z=z_coords,
-            i=faces_i, j=faces_j, k=faces_k,
-            color='rgba(220,20,60,0.6)',  # Semi-transparent crimson
-            opacity=0.75,
-            name="ModeSource",
-            hovertemplate=hovertext + "<extra></extra>",
-            # Prominent outline for visibility
-            contour=dict(
-                show=True,
-                color="darkred",
-                width=8  # Thick lines for source visibility
-            ),
-            # Flat shading for clean appearance
-            lighting=dict(
-                ambient=0.8,
-                diffuse=0.2,
-                fresnel=0.0,
-                specular=0.0,
-                roughness=1.0
-            ),
-            flatshading=True,
-            showlegend=True
-        ))
-        
-        # Add an arrow to show the propagation direction
-        self._add_direction_arrow_to_3d_plot(fig, source, vertices)
-    
-    def _add_direction_arrow_to_3d_plot(self, fig, source, plane_vertices):
-        """Add a direction arrow to show the propagation direction of the mode source."""
-        try:
-            import plotly.graph_objects as go
-        except ImportError:
-            return
-            
-        # Calculate the center of the plane
-        center = np.array([
-            sum(v[0] for v in plane_vertices) / len(plane_vertices),
-            sum(v[1] for v in plane_vertices) / len(plane_vertices),
-            sum(v[2] for v in plane_vertices) / len(plane_vertices)
-        ])
-        
-        # Determine arrow direction based on source.direction
-        arrow_length = source.wavelength * 0.8
-        
-        if source.direction == "+x":
-            arrow_end = center + np.array([arrow_length, 0, 0])
-        elif source.direction == "-x":
-            arrow_end = center + np.array([-arrow_length, 0, 0])
-        elif source.direction == "+y":
-            arrow_end = center + np.array([0, arrow_length, 0])
-        elif source.direction == "-y":
-            arrow_end = center + np.array([0, -arrow_length, 0])
-        elif source.direction == "+z":
-            arrow_end = center + np.array([0, 0, arrow_length])
-        elif source.direction == "-z":
-            arrow_end = center + np.array([0, 0, -arrow_length])
-        else:
-            # Default to +x direction
-            arrow_end = center + np.array([arrow_length, 0, 0])
-        
-        # Create arrow as a line with a cone at the end
-        fig.add_trace(go.Scatter3d(
-            x=[center[0], arrow_end[0]],
-            y=[center[1], arrow_end[1]],
-            z=[center[2], arrow_end[2]],
-            mode='lines',
-            line=dict(color='darkred', width=8),
-            name="Propagation Direction",
-            showlegend=False,
-            hoverinfo='skip'
-        ))
-        
-        # Add arrowhead as a cone
-        fig.add_trace(go.Cone(
-            x=[arrow_end[0]], y=[arrow_end[1]], z=[arrow_end[2]],
-            u=[arrow_end[0] - center[0]], v=[arrow_end[1] - center[1]], w=[arrow_end[2] - center[2]],
-            sizemode="absolute",
-            sizeref=arrow_length * 0.3,
-            colorscale=[[0, 'darkred'], [1, 'darkred']],
-            showscale=False,
-            showlegend=False,
-            hoverinfo='skip'
-        ))
-    
-    def _add_gaussian_source_to_3d_plot(self, fig, source, scale, unit):
-        """Add a Gaussian source as a sphere to the 3D plot."""
-        try:
-            import plotly.graph_objects as go
-        except ImportError:
-            return  # Skip if plotly not available
-            
-        # Create a sphere to represent the Gaussian source
-        position = source.position
-        radius = source.width * 0.5  # Use half the width as radius
-        
-        # Create sphere coordinates
-        phi = np.linspace(0, 2*np.pi, 20)
-        theta = np.linspace(0, np.pi, 20)
-        phi, theta = np.meshgrid(phi, theta)
-        
-        x = position[0] + radius * np.sin(theta) * np.cos(phi)
-        y = position[1] + radius * np.sin(theta) * np.sin(phi)
-        z = position[2] + radius * np.cos(theta)
-        
-        # Create hover text
-        hovertext = f"GaussianSource"
-        hovertext += f"<br>Position: ({position[0]*scale:.2f}, {position[1]*scale:.2f}, {position[2]*scale:.2f}) {unit}"
-        hovertext += f"<br>Width: {source.width*scale:.2f} {unit}"
-        
-        # Add the source as a semi-transparent sphere
-        fig.add_trace(go.Surface(
-            x=x, y=y, z=z,
-            colorscale=[[0, 'rgba(255,69,0,0.7)'], [1, 'rgba(255,69,0,0.7)']],  # Orange-red
-            opacity=0.7,
-            name="GaussianSource",
-            hovertemplate=hovertext + "<extra></extra>",
-            showscale=False,
-            showlegend=True
-        ))
-    
-    def _structure_to_3d_mesh(self, structure, depth, z_offset=0):
-        """Convert a 2D structure to 3D mesh data for plotly with robust triangulation for complex polygons."""
-        if not hasattr(structure, 'vertices') or not structure.vertices: return None
-        # Ensure depth has a valid value
-        if depth is None: depth = 0.1 * min(self.width, self.height)
-        # Get 2D projection of vertices for triangulation
-        vertices_2d = structure._vertices_2d() if hasattr(structure, '_vertices_2d') \
-                        else [(v[0], v[1]) for v in structure.vertices]
-        n_vertices = len(vertices_2d)
-        if n_vertices < 3: return None  # Need at least 3 vertices for a face
-        
-        # Get the actual z position from the structure
-        actual_z = z_offset
-        if hasattr(structure, 'z') and structure.z is not None: actual_z = structure.z
-        elif hasattr(structure, 'position') and len(structure.position) > 2: actual_z = structure.position[2]
-            
-        # Handle polygons with holes (like Ring structures)
-        interior_paths = getattr(structure, 'interiors', [])
-        # For complex polygons with holes, use a different approach
-        if interior_paths and len(interior_paths) > 0:
-            return self._triangulate_polygon_with_holes(vertices_2d, interior_paths, depth, actual_z)
-    
-        # For complex unified polygons, we need robust triangulation
-        try:
-            triangles = self._robust_triangulation(vertices_2d)
-        except Exception as e:
-            print(f"Warning: Robust triangulation failed ({e}), using fallback")
-            triangles = self._fallback_triangulation(vertices_2d)
-        
-        if not triangles:
-            print("Warning: All triangulation methods failed, skipping structure")
-            return None
-        
-        # Create 3D vertices by extruding the 2D shape
-        vertices_3d = []
-        # Bottom face vertices (z = actual_z)
-        for x, y in vertices_2d: vertices_3d.append([x, y, actual_z])
-        # Top face vertices (z = actual_z + depth) - maintain exact same x,y coordinates
-        for x, y in vertices_2d: vertices_3d.append([x, y, actual_z + depth])
-            
-        # Extract coordinates
-        x_coords = [v[0] for v in vertices_3d]
-        y_coords = [v[1] for v in vertices_3d]
-        z_coords = [v[2] for v in vertices_3d]
-        
-        # Create triangular faces for the mesh
-        faces_i, faces_j, faces_k = [], [], []
-        
-        # Bottom face triangulation (CCW when viewed from below = normal pointing down)
-        for tri in triangles:
-            faces_i.append(tri[0])
-            faces_j.append(tri[2])  # Swap j,k for opposite winding
-            faces_k.append(tri[1])
-                
-        # Top face triangulation (CCW when viewed from above = normal pointing up)
-        for tri in triangles:
-            faces_i.append(tri[0] + n_vertices)
-            faces_j.append(tri[1] + n_vertices)
-            faces_k.append(tri[2] + n_vertices)
-                
-        # Side faces - create vertical faces between corresponding vertices
-        for i in range(n_vertices):
-            next_i = (i + 1) % n_vertices
-            # Each side face is two triangles forming a rectangle
-            # Triangle 1: bottom-left, bottom-right, top-right
-            faces_i.append(i)
-            faces_j.append(next_i)
-            faces_k.append(next_i + n_vertices)
-            # Triangle 2: bottom-left, top-right, top-left
-            faces_i.append(i)
-            faces_j.append(next_i + n_vertices)
-            faces_k.append(i + n_vertices)
-        
-        return {
-            'vertices': (x_coords, y_coords, z_coords),
-            'faces': (faces_i, faces_j, faces_k)
-        }
-    
-    def _robust_triangulation(self, vertices_2d):
-        """Robust triangulation for complex polygons including unified shapes."""
-        if len(vertices_2d) < 3: return []
-        if len(vertices_2d) == 3: return [(0, 1, 2)]
-        if len(vertices_2d) == 4: return [(0, 1, 2), (0, 2, 3)]
-        
-        try:
-            import scipy.spatial
-            # Convert vertices to numpy array
-            points = np.array(vertices_2d)
-            # Create Delaunay triangulation
-            tri = scipy.spatial.Delaunay(points)
-            # Filter triangles to only include those inside the polygon
-            valid_triangles = []
-            for triangle in tri.simplices:
-                # Calculate triangle centroid
-                centroid = np.mean(points[triangle], axis=0)
-                # Only include triangles whose centroids are inside the polygon
-                if self._point_in_polygon_2d(centroid[0], centroid[1], vertices_2d):
-                    # Ensure CCW orientation
-                    v1 = points[triangle[1]] - points[triangle[0]]
-                    v2 = points[triangle[2]] - points[triangle[0]]
-                    if np.cross(v1, v2) > 0: valid_triangles.append(tuple(triangle))
-                    else: valid_triangles.append((triangle[0], triangle[2], triangle[1]))
-            return valid_triangles
-        except ImportError:
-            print("scipy not available, using ear clipping")
-            return self._ear_clipping_triangulation(vertices_2d)
-    
-    def _ear_clipping_triangulation(self, vertices):
-        """Ear clipping algorithm for triangulating simple polygons."""
-        if len(vertices) < 3: return []
-        
-        def is_ear(i, j, k, vertices, indices):
-            """Check if vertex j forms an ear with vertices i and k."""
-            # Get the three points
-            a = np.array(vertices[indices[i]])
-            b = np.array(vertices[indices[j]])
-            c = np.array(vertices[indices[k]])
-            
-            # Check if the angle is convex
-            ab = b - a
-            cb = b - c
-            cross = np.cross(ab, cb)
-            if cross <= 0: return False
-                
-            # Check if any other vertex is inside this triangle
-            triangle = [a, b, c]
-            for m in range(len(indices)):
-                if m not in [i, j, k]:
-                    p = np.array(vertices[indices[m]])
-                    if self._point_in_triangle(p, a, b, c): return False
-            return True
-        
-        # Initialize the indices list
-        indices = list(range(len(vertices)))
-        triangles = []
-        
-        # Continue until we can't find any more ears
-        while len(indices) > 3:
-            n = len(indices)
-            ear_found = False
-            # Look for an ear
-            for j in range(n):
-                i = (j - 1) % n
-                k = (j + 1) % n
-                if is_ear(i, j, k, vertices, indices):
-                    # Add the ear triangle
-                    triangles.append((indices[i], indices[j], indices[k]))
-                    # Remove the ear tip
-                    indices.pop(j)
-                    ear_found = True
-                    break
-            # If no ear is found, fall back to simple triangulation
-            if not ear_found: break
-        
-        # Add the final triangle
-        if len(indices) == 3: triangles.append((indices[0], indices[1], indices[2]))
-        
-        return triangles
-    
-    def _fallback_triangulation(self, vertices_2d):
-        """Simple fallback triangulation methods."""
-        if len(vertices_2d) < 3: return []
-        if len(vertices_2d) == 3: return [(0, 1, 2)]
-        if len(vertices_2d) == 4: return [(0, 1, 2), (0, 2, 3)]
-        
-        # Try convex hull based triangulation
-        try: return self._convex_hull_triangulation(vertices_2d)
-        except Exception:
-            # Last resort: simple fan triangulation from centroid
-            print("Warning: Using simple fan triangulation as last resort")
-            triangles = []
-            for i in range(1, len(vertices_2d) - 1): triangles.append((0, i, i + 1))
-            return triangles
-    
-    def _convex_hull_triangulation(self, vertices_2d):
-        """Triangulation using convex hull approach."""
-        try:
-            import scipy.spatial
-            points = np.array(vertices_2d)
-            # Compute convex hull
-            hull = scipy.spatial.ConvexHull(points)
-            hull_vertices = hull.vertices
-            # If the polygon is already convex, use simple fan triangulation
-            if len(hull_vertices) == len(vertices_2d):
-                triangles = []
-                for i in range(1, len(vertices_2d) - 1): triangles.append((0, i, i + 1))
-                return triangles
-            # For non-convex polygons, try to decompose
-            else: return self._decompose_polygon(vertices_2d)
-        except ImportError: raise Exception("scipy not available for convex hull triangulation")
-    
-    def _decompose_polygon(self, vertices_2d):
-        """Decompose complex polygon into simpler triangles."""
-        # Simple polygon decomposition - find the "most convex" triangulation
-        triangles = []
-        n = len(vertices_2d)
-        # Use a modified fan triangulation that skips problematic triangles
-        center_idx = 0
-        for i in range(1, n - 1):
-            next_i = i + 1
-            # Check if this triangle is valid (not self-intersecting)
-            if self._is_valid_triangle(vertices_2d, center_idx, i, next_i): triangles.append((center_idx, i, next_i))
-        
-        return triangles if triangles else [(0, 1, 2)]  # Emergency fallback
-    
-    def _is_valid_triangle(self, vertices, i, j, k):
-        """Check if a triangle formed by vertices i, j, k is valid."""
-        # Check for degenerate triangles (zero area)
-        p1, p2, p3 = vertices[i], vertices[j], vertices[k]
-        area = abs((p2[0] - p1[0]) * (p3[1] - p1[1]) - (p3[0] - p1[0]) * (p2[1] - p1[1]))
-        return area > 1e-10  # Small threshold for numerical stability
-    
-    def _point_in_polygon_2d(self, x, y, polygon_vertices):
-        """Simple point-in-polygon test for 2D coordinates."""
-        n = len(polygon_vertices)
-        inside = False
-        p1x, p1y = polygon_vertices[0]
-        for i in range(1, n + 1):
-            p2x, p2y = polygon_vertices[i % n]
-            if y > min(p1y, p2y):
-                if y <= max(p1y, p2y):
-                    if x <= max(p1x, p2x):
-                        if p1y != p2y:
-                            xinters = (y - p1y) * (p2x - p1x) / (p2y - p1y) + p1x
-                        if p1x == p2x or x <= xinters:
-                            inside = not inside
-            p1x, p1y = p2x, p2y
-        return inside
-
-    def _triangulate_polygon_with_holes(self, exterior_vertices, interior_paths, depth, z_offset):
-        """Handle polygons with holes (like Ring structures) using proper triangulation."""
-        try:
-            # Try to use a proper constrained triangulation library if available
-            import scipy.spatial
-            from matplotlib.path import Path
-            # For now, use a simpler approach for Ring-like structures
-            # Create separate meshes for exterior and interior, then combine
-            n_ext = len(exterior_vertices)
-            total_vertices = n_ext
-            all_vertices_2d = list(exterior_vertices)
-            # Add interior vertices (these are now already 2D)
-            interior_starts = []
-            for interior in interior_paths:
-                interior_starts.append(total_vertices)
-                all_vertices_2d.extend(interior)
-                total_vertices += len(interior)
-            # Create 3D vertices
-            vertices_3d = []
-            # Bottom face vertices
-            for x, y in all_vertices_2d: vertices_3d.append([x, y, z_offset])
-            # Top face vertices
-            for x, y in all_vertices_2d: vertices_3d.append([x, y, z_offset + depth])
-            # Extract coordinates
-            x_coords = [v[0] for v in vertices_3d]
-            y_coords = [v[1] for v in vertices_3d]
-            z_coords = [v[2] for v in vertices_3d]
-            faces_i, faces_j, faces_k = [], [], []
-            # For Ring structures, create triangular strip between inner and outer
-            if len(interior_paths) == 1 and len(interior_paths[0]) == len(exterior_vertices):
-                # This is likely a Ring structure with equal number of points
-                inner_start = interior_starts[0]
-                # Create triangular strips connecting outer to inner
-                for i in range(n_ext):
-                    next_i = (i + 1) % n_ext
-                    # Bottom face triangles (outer ring)
-                    outer_i = i
-                    outer_next = next_i
-                    inner_i = inner_start + i
-                    inner_next = inner_start + next_i
-                    # Triangle 1: outer_i -> outer_next -> inner_i
-                    faces_i.append(outer_i)
-                    faces_j.append(outer_next)
-                    faces_k.append(inner_i)
-                    # Triangle 2: outer_next -> inner_next -> inner_i  
-                    faces_i.append(outer_next)
-                    faces_j.append(inner_next)
-                    faces_k.append(inner_i)
-                    # Top face triangles (same pattern but offset by total_vertices)
-                    top_offset = total_vertices
-                    # Triangle 1: outer_i -> inner_i -> outer_next (reversed winding)
-                    faces_i.append(outer_i + top_offset)
-                    faces_j.append(inner_i + top_offset)
-                    faces_k.append(outer_next + top_offset)
-                    # Triangle 2: outer_next -> inner_i -> inner_next (reversed winding)
-                    faces_i.append(outer_next + top_offset)
-                    faces_j.append(inner_i + top_offset)
-                    faces_k.append(inner_next + top_offset)
-                # Side faces for outer ring
-                for i in range(n_ext):
-                    next_i = (i + 1) % n_ext
-                    # Outer side face
-                    faces_i.append(i)
-                    faces_j.append(next_i)
-                    faces_k.append(i + total_vertices)
-                    faces_i.append(next_i)
-                    faces_j.append(next_i + total_vertices)
-                    faces_k.append(i + total_vertices)
-                # Side faces for inner ring (reversed winding for inward-facing)
-                for i in range(len(interior_paths[0])):
-                    next_i = (i + 1) % len(interior_paths[0])
-                    inner_i = inner_start + i
-                    inner_next = inner_start + next_i
-                    # Inner side face (reversed winding)
-                    faces_i.append(inner_i + total_vertices)
-                    faces_j.append(inner_next + total_vertices)
-                    faces_k.append(inner_i)
-                    faces_i.append(inner_i)
-                    faces_j.append(inner_next + total_vertices)
-                    faces_k.append(inner_next)
-            return {
-                'vertices': (x_coords, y_coords, z_coords),
-                'faces': (faces_i, faces_j, faces_k)
-            }
-            
-        except Exception as e:
-            # Fallback to simple approach if triangulation fails
-            print(f"Warning: Complex triangulation failed, using simple approach: {e}")
-            return self._structure_to_3d_mesh_simple(exterior_vertices, depth, z_offset)
-    
-    def _structure_to_3d_mesh_simple(self, vertices_2d, depth, z_offset):
-        """Fallback simple mesh generation."""
-        n_vertices = len(vertices_2d)
-        # Create 3D vertices (vertices_2d is already in 2D format)
-        vertices_3d = []
-        for x, y in vertices_2d: vertices_3d.append([x, y, z_offset])
-        for x, y in vertices_2d: vertices_3d.append([x, y, z_offset + depth])
-        x_coords = [v[0] for v in vertices_3d]
-        y_coords = [v[1] for v in vertices_3d]
-        z_coords = [v[2] for v in vertices_3d]
-        faces_i, faces_j, faces_k = [], [], []
-        # Simple fan triangulation for bottom
-        for i in range(1, n_vertices - 1):
-            faces_i.append(0)
-            faces_j.append(i)
-            faces_k.append(i + 1)
-        # Simple fan triangulation for top (reversed)
-        for i in range(1, n_vertices - 1):
-            faces_i.append(n_vertices)
-            faces_j.append(n_vertices + i + 1)
-            faces_k.append(n_vertices + i)
-        # Side faces
-        for i in range(n_vertices):
-            next_i = (i + 1) % n_vertices
-            faces_i.append(i)
-            faces_j.append(next_i)
-            faces_k.append(i + n_vertices)
-            faces_i.append(next_i)
-            faces_j.append(next_i + n_vertices)
-            faces_k.append(i + n_vertices)
-        return {
-            'vertices': (x_coords, y_coords, z_coords),
-            'faces': (faces_i, faces_j, faces_k)
-        }
-    
-    def _point_in_triangle(self, point, a, b, c):
-        """Check if a point is inside a triangle using barycentric coordinates."""
-        x, y = point
-        x1, y1 = a
-        x2, y2 = b
-        x3, y3 = c
-        denominator = (y2 - y3) * (x1 - x3) + (x3 - x2) * (y1 - y3)
-        if abs(denominator) < 1e-10: return False
-        a_coord = ((y2 - y3) * (x - x3) + (x3 - x2) * (y - y3)) / denominator
-        b_coord = ((y3 - y1) * (x - x3) + (x1 - x3) * (y - y3)) / denominator
-        c_coord = 1 - a_coord - b_coord
-        return a_coord >= 0 and b_coord >= 0 and c_coord >= 0
+        """Display the design using 3D plotly (delegated to beamz.viz)."""
+        return viz.show_design_3d(self, unify_structures, max_vertices_for_unification)
 
     def __str__(self):
         return f"Design with {len(self.structures)} structures ({'3D' if self.is_3d else '2D'})"
